Restructured files in components folder and fixed some bugs creating console log errors in challenge details section
