+++ client/src/Components/AppComponents/Auth/NoMatch/index.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { Col, Row, Container } from "../../UiComponents/Grid";
+import { Col, Row, Container } from "../../../BootstrapComponents/Grid";
 // import Jumbotron from "../../UiComponents/Jumbotron";
 
 function NoMatch() {
