+++ client/src/Components/AppComponents/Profile/ChallengeDetails/ChallengeDetails.css
@@ -1,0 +1,31 @@
+.challenge-details-card{
+    margin:auto;
+    margin-top:5%;
+    width:80%;
+    padding:3%;
+    color:whitesmoke;
+  }
+.challenge-details-card .day-section{  
+  width:14%;
+  height:7em;
+  float:left;
+  border: 2px solid #D0E4F5;
+  border-right:none;
+  padding:0.2%;
+  margin:0px;
+  border-bottom:none;
+}
+.challenge-details-card .day-section:nth-child(7) , .day-section:nth-child(14),.day-section:nth-child(21),.day-section:nth-child(28),.day-section:nth-child(30){
+  border-right: 2px solid #D0E4F5;
+}
+.challenge-details-card .day-section:nth-child(24),.day-section:nth-child(25),.day-section:nth-child(26),.day-section:nth-child(27),.day-section:nth-child(28),.day-section:nth-child(29),.day-section:nth-child(30){
+  border-bottom: 2px solid #D0E4F5;
+}
+.day-section:hover{
+  background-color:grey;
+}
+.details-image{
+  float:left;
+  width:30%;
+  padding:105;
+}
