+++ client/src/Components/AppComponents/Profile/ProfileNav/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 import "./style.css";
-import { FormBtn,FormBtnlink } from "../../UiComponents/Form";
+import { FormBtn,FormBtnlink } from "../../../BootstrapComponents/Form";
 import {Link} from "react-router-dom";
 
 function ProfileNav({ userName, profilePic,bio }) {
