--- client/src/Components/Pages/ChallengeDetails/ChallengeDetails.css
@@ -1,31 +1,0 @@
-.challenge-details-card{
-    margin:auto;
-    margin-top:5%;
-    width:80%;
-    padding:3%;
-    color:whitesmoke;
-  }
-.challenge-details-card p{  
-  width:14%;
-  height:7em;
-  float:left;
-  border: 2px solid #D0E4F5;
-  border-right:none;
-  padding:0.2%;
-  margin:0px;
-  border-bottom:none;
-}
-.challenge-details-card p:nth-child(7) , p:nth-child(14),p:nth-child(21),p:nth-child(28),p:nth-child(30){
-  border-right: 2px solid #D0E4F5;
-}
-.challenge-details-card p:nth-child(24),p:nth-child(25),p:nth-child(26),p:nth-child(27),p:nth-child(28),p:nth-child(29),p:nth-child(30){
-  border-bottom: 2px solid #D0E4F5;
-}
-p:hover{
-  background-color:grey;
-}
-.details-image{
-  float:left;
-  width:30%;
-  padding:105;
-}
