--- client/src/Components/Pages/DoneChallenges/DoneChallenges.css
@@ -1,86 +1,0 @@
-.browse-card{
-    height:100%;
-    background-color: rgb(106, 103, 99);
-    padding:0px;
-}
-.card-img-top{
-    height:13em;
-}
-.challenge-title{
-font-size: 2em;
-color:whitesmoke;
-}
-.challenge-text{
-    color:whitesmoke;
-    font-size:1em;
- }
-
-
-.challenge-info{
-    background-color:white;
-    padding:5%;
-}
-.fa-share-alt{
-    float:right;
-}
-.card-body{
-    text-align: center
-}
-
-.card {
-    background: rgb(108, 108, 108);
-    padding:2px;
-    border-radius: 2px;
-    margin: 1rem;
-    position: relative;
-    width: 30%;
-    height:50%;
-    box-shadow: 0 0.5px 2px #999, 0 1px 1px #999;
-    text-align: left;
-  }
-  body{
-    background-color:rgb(67, 66, 66);
-    }
-  
-  .card > .img-container {
-    height: 100%;
-    width:30%;
-    float:left;
-    text-align: center;
-   
-  }
-  .card > .img-container>h4{
-    color:white;
-  }
-  
-  
-  .card > .img-container > img {
-    width: 100%
-  }
-  
-  .card > .content {
-    padding-left: 1rem;
-    padding-right: 1rem;
-    font-size: 15px;
-  }
-  
-  .card > .content > ul {
-    list-style-type: none;
-  }
-  
-  .card > .content > ul > li {
-    margin-top: 18px;
-  }
-  
-  .card > .remove {
-    color: #ff1744;
-    font-size: 48px;
-    position: absolute;
-    top: -20px;
-    right: 5px;
-    cursor: pointer;
-  }
-  
-  .card > .remove:hover {
-    color: #d50000;
-  }
